Deleted legacy openvino API dependency and modularized code
1. Modularized prepared model code by moving prepared model independant
functions to utils
2. Created separate CPU, GNA folders and respective prepared model
classes
3. Removed dependency of legacy openvino API's in HAL code
4. Added basepreparedmodel, utils, cpu & gpu preparedmodel to Android.bp
5. Changed VLOG's to ALOG's
6. removed mCreateNgraph code
7. removed ngraph checker in Preparedmodel
8. Deleted Android.bp in graphAPI
9. Further cleanedup IENetworks.h
10. Renamed member variable enginePtr to mEnginePtr
