+++ Driver.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef ANDROID_ML_NN_VPU_DRIVER_H
-#define ANDROID_ML_NN_VPU_DRIVER_H
+#ifndef ANDROID_ML_NN_DRIVER_H
+#define ANDROID_ML_NN_DRIVER_H
 #include <android/hardware/neuralnetworks/1.0/IDevice.h>
 #include <android/hardware/neuralnetworks/1.0/IExecutionCallback.h>
 #include <android/hardware/neuralnetworks/1.0/IPreparedModel.h>
@@ -28,6 +28,7 @@
 #include <android/hardware/neuralnetworks/1.2/IPreparedModel.h>
 #include <android/hardware/neuralnetworks/1.2/IPreparedModelCallback.h>
 #include <android/hardware/neuralnetworks/1.2/types.h>
+
 #include <string>
 
 namespace android {
@@ -35,24 +36,28 @@ namespace hardware {
 namespace neuralnetworks {
 namespace nnhal {
 
+// For HAL-1.0 version
+using namespace ::android::hardware::neuralnetworks::V1_0;
+using V1_0_Model = ::android::hardware::neuralnetworks::V1_0::Model;
+using V1_0_Operation = ::android::hardware::neuralnetworks::V1_0::Operation;
+using V1_0_Capabilities = ::android::hardware::neuralnetworks::V1_0::Capabilities;
+
+// For HAL-1.1 version
+using namespace ::android::hardware::neuralnetworks::V1_1;
+using V1_1_Model = ::android::hardware::neuralnetworks::V1_1::Model;
+using V1_1_Operation = ::android::hardware::neuralnetworks::V1_1::Operation;
+using V1_1_Capabilities = ::android::hardware::neuralnetworks::V1_1::Capabilities;
+
+// For HAL-1.2 version
+using namespace ::android::hardware::neuralnetworks::V1_2;
 using Model = ::android::hardware::neuralnetworks::V1_2::Model;
 using Operand = ::android::hardware::neuralnetworks::V1_2::Operand;
-using V1_0_Model = ::android::hardware::neuralnetworks::V1_0::Model;
-using V1_1_Model = ::android::hardware::neuralnetworks::V1_1::Model;
 using Operation = ::android::hardware::neuralnetworks::V1_2::Operation;
-using V1_0_Operation = ::android::hardware::neuralnetworks::V1_0::Operation;
-using V1_1_Operation = ::android::hardware::neuralnetworks::V1_1::Operation;
 using OperationType = ::android::hardware::neuralnetworks::V1_2::OperationType;
 using OperandType = ::android::hardware::neuralnetworks::V1_2::OperandType;
 using Capabilities = ::android::hardware::neuralnetworks::V1_2::Capabilities;
-using V1_0_Capabilities = ::android::hardware::neuralnetworks::V1_0::Capabilities;
-using V1_1_Capabilities = ::android::hardware::neuralnetworks::V1_1::Capabilities;
 
-using namespace ::android::hardware::neuralnetworks::V1_1;
-using namespace ::android::hardware::neuralnetworks::V1_0;
-using namespace ::android::hardware::neuralnetworks::V1_2;
 using ::android::hardware::MQDescriptorSync;
-
 using HidlToken = android::hardware::hidl_array<uint8_t, 32>;
 
 // Base class used to create vpu drivers for the NN HAL.  This class
@@ -63,22 +68,28 @@ using HidlToken = android::hardware::hidl_array<uint8_t, 32>;
 class Driver : public ::android::hardware::neuralnetworks::V1_2::IDevice {
 public:
     Driver() {}
-    Driver(const char* name) : mName(name) {}
+    Driver(const char* name) : mDeviceName(name) {}
 
     ~Driver() override {}
+
+    // For HAL-1.0 version
     Return<void> getCapabilities(getCapabilities_cb cb) override;
-    Return<void> getCapabilities_1_1(getCapabilities_1_1_cb cb) override;
-    Return<void> getCapabilities_1_2(getCapabilities_1_2_cb cb) override;
     Return<void> getSupportedOperations(const V1_0_Model& model,
                                         getSupportedOperations_cb cb) override;
-    Return<void> getSupportedOperations_1_1(const V1_1_Model& model,
-                                            getSupportedOperations_1_1_cb cb) override;
-    Return<void> getSupportedOperations_1_2(const Model& model,
-                                            getSupportedOperations_1_2_cb cb) override;
     Return<ErrorStatus> prepareModel(const V1_0_Model& model,
                                      const sp<V1_0::IPreparedModelCallback>& callback) override;
+
+    // For HAL-1.1 version
+    Return<void> getCapabilities_1_1(getCapabilities_1_1_cb cb) override;
+    Return<void> getSupportedOperations_1_1(const V1_1_Model& model,
+                                            getSupportedOperations_1_1_cb cb) override;
     Return<ErrorStatus> prepareModel_1_1(const V1_1_Model& model, ExecutionPreference preference,
                                          const sp<V1_0::IPreparedModelCallback>& callback) override;
+
+    // For HAL-1.2 version
+    Return<void> getCapabilities_1_2(getCapabilities_1_2_cb cb) override;
+    Return<void> getSupportedOperations_1_2(const Model& model,
+                                            getSupportedOperations_1_2_cb cb) override;
     Return<ErrorStatus> prepareModel_1_2(const Model& model, ExecutionPreference preference,
                                          const hidl_vec<hidl_handle>& modelCache,
                                          const hidl_vec<hidl_handle>& dataCache,
@@ -94,7 +105,7 @@ public:
     Return<void> getNumberOfCacheFilesNeeded(getNumberOfCacheFilesNeeded_cb cb) override;
 
 protected:
-    std::string mName;
+    std::string mDeviceName;
 };
 
 }  // namespace nnhal
@@ -102,4 +113,4 @@ protected:
 }  // namespace hardware
 }  // namespace android
 
-#endif  // ANDROID_ML_NN_VPU_DRIVER_H
+#endif  // ANDROID_ML_NN_DRIVER_H
